Retire FileAttachmentUploadService and refactor FileAttachmentBlobService
This removes the FileAttachmentUploadService and instead handles the
concerns of that in a create_initial method on FileAttachment::Revision.
This is to reflect that FileAttachmentUploadService is not performing
any uploading and thus it's name had become somewhat irrelevant.

It refactors the FileAttachmentBlobService to act more like a service
that performs an action and returns a result, as opposed to a model
where various attributes are looked up.

It was a total pain thinking about the attributes and names for this
service. This is likely because the concept this object is dealing with
is quite odd: it takes an uploaded file, saves it to
database/persistent storage and then returns the necessary attributes
that can be persisted in a BlobRevision record, rather than an actual
record.

It'd probably be nicer if this could return a BlobRevision directly but
this has a couple of problems to resolve with the updaters and would
need more parameters.