Initial modelling for access limit
This creates an access limit model that is associated with an edition.

The access limit model stores:

- limit_type: whether all organisations or just primary organisation
  apply
- edition: the edition the limit applies to
- revision_at_creation: which revision the edition represented at the
  point of access limiting to know which changes were limited and not
- created_by: the user who created it
- created_at: when it was created
- active: whether the limit applies

The intention is that for any change to an access limit a new model is
created where the active flag is changed on the one that is active. When
removing an access limit the active flag is changed to false.

A tricky decision was deciding the best way to associate this with an
edition. The option I rejected was storing access limit as an optional
foreign key on an edition, as this would mean that we would have a
bi-directional association between edition and access limit. This makes
it challenging to create models and association methods can go to two
places. This is the approach we've taken for a status.

The approach I did take was similar to the one taken with a Document and
Edition relationship where we store a flag on the edition as to whether
it is live or current. This uses a flag of active to indicate that the
access limit applies or not. A unique index on the active flag and
edition_id prevents there being two access limits applying at the same
time.

The reason we can do this for AccessLimit rather than Status is because
AccessLimit is an optional association. Whereas this approach can't be
used for a mandatory one as there is nothing to force it's presence.

By having an active flag that changes this model cannot be readonly, to
counter this all the attributes that should not be edited have been set
to `attr_readonly`. Using this stops users editing but silently rather
than an exception, so it's less favourable than using readonly.