Make access limit edition query more efficient
Previously we were retrieving unnecessary data in order to check the
user against the access limit for an edition. This refactors the
Edition#find_current method in order to expose the mechanism to find an
edition using the standard content_id:locale path param, which means we
can re-use this code but modify the data that's retrieved.

Unfortunately we still end up doing extra queries for the edition orgs
when an access limit is present, since our current associations with
access limits do not support caching of editions via `inverse_of`. This
seems reasonable, as we expect most editions won't be access limited.