Update the instance values of document topics
Memoizes the value of `document_topics` so that they don't need to be fetched from publishing-api
every time they are accessed.

As `topic_content_ids` is being updated before the call to publishing_api, we no longer need to pass these
values into the leaf and legacy method, but we do need to remember what the value of the unknown taxons
were before the user made any changes, as we need to be sure that we're not overwriting them.