Fixed some of the errors
Lots of the tests were failing due to not being able to access document_type in the normal way. This has been delegated through on all of the models down to the metadata revsion. Some of the tests are still failing because there are no editions associated with the documents that they use.