Refactor setting the status
We need to be able to set two Status objects for the withdrawn edition, the first with details about
when the edition was published and the second for when it was withdrawn. This is to allow Content Publisher
to unwithdraw the edition if needed.

Refactoring setting that status and giving it a more meaningful name, will allow a more status entries to be added
later.