Use MigrateState and move status logic to CreateEdition
Switches the code in CreateEdition to use the new
MigrateState class and pulls out some logic around creating a status
from the removed CreateStatus class.

This new layout of code will make it easier to introduce the
unpublishing logic in later commits and gets rid of the CreateStatus
class as it contained mixed responsibilities and didn't actually
'create' a status.