Updates FileAttachmentBlobService
Changes the behaviour of the FileAttachmentBlobService slightly by
moving the creating of a unique file name to the interactors. This has
the advantage of reducing the amount of params needed to instantiate the
class and will allow us to add file_attachments to the import much more
easily.

One of the tests which has been removed ("allows keeping the name if
this is a replacement for the attachment with the same name") is already
covered by the /replace_file_attachment_file_spec.rb feature and the
other one is just basically testing the UniqueFilenameService.