Create WhitehallMigration::AssetImport for Image::Revision
I've added a default "s960" variant of images to the image factory
so that this behaviour can be tested.

I've also had to specify an 'assets' association on DocumentImport
so that I can properly test the new behaviour. This 'assets'
association will come in handy later when we write the migration
logic.