Create WhitehallMigration::AssetImport factory
This factory is about to be used by a number of subsequent
commits. For ease of rebasing/editing, I've encapsulated it in
one commit.

For the record, I attempted to name the factories `:image` and
`:file_attachment`, but as these names clashed with existing
factory names, it was trying to set `image`/`file_attachment`
attributes instead. I've gone into more detail here:
https://github.com/thoughtbot/factory_bot/issues/528#issuecomment-569951414