Create WhitehallMigration::AssetImport for FileAttachment::Revision
I've had to make the image_revision/file_attachment_revision
associations optional as only one of them should exist per asset.
In a future commit I intend to write a validator to ensure this is
the case.

As JSON doesn't support the concept of `nil`, and we don't
anticipate encountering a variant of an asset that contains nils,
I've tweaked the factory data for file attachments so that a
content type and url are actually defined.