Prevent importing the same assets multiple times
Images/attachments are allowed to have the same filenames - we
increment the name by 1 every time - but the preceding prefix
has to be different. If an identical URL is encountered, we've
already processed the asset once and don't need to do it again.

However, we do need to return the processed asset even if we've
processed it in the past, otherwise the imported document will
be missing the image/fileattachment revision. So this has become
a 'find or create' method rather than simply 'create'.