Get a list of documents to import from Whitehall
Updates the import:whitehall_migration rake task to get a list of
relevant documents and creates a WhitehallMigration::DocumentImport
record (in a pending state) for each document.

The complexity here is due to Whitehall paginating the results.  We loop
through each page until we find a page has fewer than the expected
number of results.