Withdraw render error rather than redirect
This allows us to return the appropriate status code of 503 rather than
having it look like everything is ok with a 302 response.

It turns out that using a flash as a hash produces different key types
depending on whether they're part of a redirect or not (symbols without,
strings with) presumably due to JSON. This changes the layout to be
resilient to both.