Prevent migrating assets that have already been processed
If a migration fails, we need to be able to re-rerun the migration
task. We don't want to run into issues needlessly re-migrating
assets that have already been processed.

We had hoped to use Rails 'scopes' for this but it was far more
complex than originally thought, as we need to scope on an association
rather than a static method call. There is also some complexity to
do with factories. See:
https://github.com/alphagov/content-publisher/pull/1568#discussion_r362206962

As a result, I've opted for a `migratable_assets` method on the
DocumentImport class which returns a subset of its `assets`.