Simplify the MigrateAssets tests
This was prompted by a desire to see if we could avoid the need for
distinct file_attachment_asset and image_asset factories. These are
somewhat undesirable as these models are created automatically by
Image and FileAttachment BlobRevisions classes so it was an unequal
relationship to sometimes creating these manually.

I then got a bit carried away making amends to the test - I'm sorry 😢

I thought that it made sense that the whitehall_migration_asset_import factory
worked with live versions of assets since this would be the happy path
behaviour for migrating an asset. This allowed simplifying a variety of
tests.

I removed the global variables of asset, asset_url and
original_asset_url. For asset this was because we already assigned our
own asset in lots of tests, for asset_url this was because we can look
this up quite quickly on the asset_import model we build and with
original_asset_url I couldn't find usages when we assert on it - so
maybe I refactored the need away.

These changes do reduce the amount of explicit information set, but I
think this is ok since the models that provide this data are not the
test subject.