Expand integrity problem description
The description of the failing check only mentions which has failed. To
make it slightly easier to debug the content of the conflicting field
have been added to the problem list

The body text has been excluded from this as the field contents are too
log to add to an error string.