Add migration to store integrity check data
Add fields to store the list of problems for a failed integrity check,
and the proposed payload that the check was made against.

The proposed payload needs to be stored, because it is created during a
transaction and rollback when the transaction fails. The payload is
difficult to recreate as it relies on the document existing as well as
all of the editions and revisions.

A copy of the export taken from Whitehall is already being store in the
whitehall_migration_document_imports table so this seemed like the best
place to also keep the proposed payload.

We don't need to store what publishing-api expects because it's easy
enough to see that by calling the get_content, get_live_content and
get_links api methods.