Rename services to match a verb/noun pattern
This renames a few services to follow a pattern where the single action
a service performs is described in terms of a verb, like 'create' or
'publish', operating on a noun, like 'filename' or 'scheduled
publishing'. In the case of 'edition' vs 'draft', I've gone with
'draft', which is less ambiguous than an arbitrary edition.