Reload based approach to atomic imports
Our problem is that we mutate the document_import object during the
import, which conflicts with our 'all or nothing' strategy for
transactions. What happens is that we create a document and various
other models in the database and then, when a transaction fails, roll
them back. However as we set the document attribute on a document_import
the rolled back model still exists in memory and when we save the
document_import it tries to save the document model meaning we at best
get an exception and at worst get a half created document.

An approach to resolve this is to reload the document_import object at
the point of the exception occurring.

A different, more complex context based approach, is considered in
https://github.com/alphagov/content-publisher/compare/atomic-imports?expand=1

Couple of other tweaks this work did:

- Move setting a document as imported to be in the transaction. That way, if
  this step raises an exception we haven't persisted a document but failed to
set the status.
- We no longer need `WhitehallImporter::Import#call` to return a document to
  `WhitehallImporter`, since assigning the document to the document import is
no longer handled there. This change caused some tests to fail as they were
relying on receiving the document from `WhitehallImporter::Import#call` to
check various values on the document. These have been updated.