Extract getting updater params into Field classes
This adds a new method to each Field class to co-locate the processing
of params with the Field that makes them present in the first place.

The name 'updater_params' is meant to be specific to the single use case
of generating a hash for updating a revision. Although our downstream
code may support both symbol and string keys (e.g. from params#permit),
having a more deterministic, symbol-only return value is easier to test.