Discard Document Import Job on Standard Error
We've decided not to retry the Whitehall Document Import Job when a
StandardError is raised; instead we want to persist the error and update the
document import with the appropriate state. Since this is similar behaviour to
when retries have been exhausted, we can extract this logic out to its own
method.

Now that StandardError is being handled by Whitehall Document Import Job, we
can remove the rescue from both the `#import` and `#sync` methods in the
Whitehall Importer.