Reset DocumentType return methods between tests
Every time a 'build(:document_type)' is called, a randomly
generated document type is appended to the return value of
DocumentType.all, and Supertype.all.first.document_types.
This pollutes the global state and causes surprises.

Resetting this state at the end of every specfile run isn't
particularly clean, but it's cleaner than letting the
problem get worse. I haven't had to reset the Supertype
factory pollution as we're always grabbing the 'first'
Supertype, so there are currently no side effects to adding
multiple.

The motivation for this change is so that we can assert
the number of items returned from DocumentType.all matches
the number of document types in our config.