Move exporting Whitehall document to the Import class
This moves exporting the Whitehall document to the Import class, since that's
where we'll be making use of it. This also means we won't need to do another
check that the document import is pending, and we can successfully rollback all
changes made to document import in the event that an error occurs, rather than
needing to rescue the error and reload the object as we currently do. A final
benefit is we can remove some of the remaining logic in `#import_and_sync`,
paving the way for it to be removed completely.