Convert asset_url to a string
When sending the payload to Publishing API, there is a possibility of
attachments having a nil URL (e.g. attachments that are not yet migrated
from Whitehall to Content Publisher).

This causes an error since Govspeak takes the attachment data [1] which
is put into a link_to request by govuk_publishing_components [2].  The
link_to method fails, since a URL cannot be nil.

Therefore we need to have a string, even if empty, in the payload so this
error does not occur.

A number of other approaches were considered but dismissed:
* Allow govuk_publishing_components allow a nil url: we would have to
  change the component and make it less valid
* Add an in_app argument to every payload method then utilise
  in_app_html instead of payload_html: this would mean the migration
  spills through the system into unrelated code
* Create assets on asset manager at the point of import, then discard
  them if the migration fails: we run the risk of further failure during
  the cleanup
* Create a mode that can be applied to assets during the migration to
  mark them as such and give them a placeholder URL: this would require
  changes to the model which are only relevant to migration

[1] https://github.com/alphagov/govspeak/blob/f8131efa42c2838a6c94288bd0404e13b0f17ad2/lib/govspeak.rb#L368-L372
[2] https://github.com/alphagov/govuk_publishing_components/blob/master/app/views/govuk_publishing_components/components/_attachment_link.html.erb#L33-L36