WIP - Update DocumentTypeSelectionInteractor
The main job of this interactor is to decide what the controller needs to do next.
So the controller needs to know:
1. Whether the user needs to refine their selection and select another document type
2. If the document type is managed elsewhere, and where to redirect to.
3. If a document has been created.

Rather than return a hash of these things, they have been added to the context.

If a document_type_id is returned, the controller needs to re-render the document_type_selection view and pass in that id.
If managed_elsewhere_url is returned, the controller needs to redirect to that url
If a document is returned, the controller needs to redirect to the content_path for that document

None of this has been tested, which is the next step.