Introduce alert for lack of government data
Trello - https://trello.com/c/FaHAEajH/1318-alert-2ndline-when-we-have-no-cached-data-on-governments

This will introduce an alert for 2nd line in the even that no government data is available. Content publisher gets its government data from a background task from the publishing api and whilst this shouldn't fail in the event that it does it will be hard for 2nd line to troubleshoot the issue. This commit iterates the `/heathcheck` endpoint to raise the appropriate response in the event that government data is unavailable or hasn't been updated in the past 6 hours.