Create new routes and controller actions for document type selection
Add new ones for now rather than editing the existing ones, so it's easier to follow the changes, and to make additions to the tests.
The "old" code will be removed later on.

The new view has placeholder code, of nils in place of missing internationalisation files. These will also be populated later.