Fix the import of embedded files in news articles
When migrating two of NDA's news articles from Whitehall to Content
Publisher, we encounter integrity check failures on the body.  The
cause of this is the presence of image markdown in the proposed
payload which we have failed to convert to an actual image, i.e.
we're sending '[Image:wood_640_x_960.jpg]' rather than '<img ....'.

The documents are:
ID: 431866 Content ID: 809d70b1-3e9b-4833-a28b-41366a925805
ID: 370367 Content ID: 5a88536b-7831-4184-b5fd-c40275ba7558

It is expeced that a similar situation will occur with attachments
due to the similarity of their handling.

Files are defined in GovSpeak as !!2 for images and !@2 for
attachments.  Both require two new lines to be entered prior to the
file (e.g. \n\n!!2 or \n\n!@2) in order to be correctly converted
into HTML correctly.

We need to fix the import so these files are correctly parsed,
and the proposed payload for Publishing API matches the currently
published document.