Use params to replay edit image inputs
Previously we relied on the image revision having the previously
submitted values when re-rendering the edit with requirements issues.
Using a dirty image revision lead to the undesirable outcome of
the alt text on the thumbnail being set to the potentially invalid
entry. This changes the edit page to use 'params' directly when
replaying values, in preparation for fixing this issue.