Change "Create new document" button to use new route
Also updates the link on the guidance page.

Also updates the feature tests to check for the document type selection localisation text.

The yaml values for DocumentTypeSelection are being accessed directly in the feature tests.
This is because creating a dynamic factory for a DocumentTypeSelection is time consuming.
Hardcoding the values for now, allows other pieces of work that depend on this to be unblocked.
The factory will be created in another PR.
The side-effect of this is that stubs need to be added for publishing-api so that the document can
be created.