Set published_at times
In this commit:

* Sets published_at time on publishing
* Populates both publishing times (first_published_at and published_at)
  in content-publisher before publishing to pub-api.
* Adds some tests for the above

One weirdness in this code is having to reload the document here -
first_published_at = document.reload.first_published_at. This needs to
be done as assigning first_published_at to the documents
first_published_at value doesn't actually retrieve the value from the db
as the document let is populated with the document attributes
in memory before saving to the db (to save a read and write when
creating a record and assigning it a variable). Then when comparing
times in the expectation ruby time(+6dp) is compared to postgres(6dp) time where
ruby time is more accurate leading to the comparison to fail.

Also would've used an instance variable for time but the linter shouted
at me :(