Refactor SelectInteractor to return all redirects
Now that we know what all of the possible routes the select action can
redirect users to, we can move that logic to the SelectInteractor.

This has the benefit of reducing the amount of information that the
interactor as to return, for example, the controller no longer needs to
what the `current_selection` is(i.e. the radio options for the next page
that needs to be loaded).