Update revision_synced in Interactor::Failure
We have a tricky situation here where we want to perform a DB save when
an interactor fails, whereas normally this would halt a transaction. The
approach I've taken to resolving this is to catch the
Interactor::Failure exception and then perform this DB update. The
interactor will be started marked as a failure despite this exception
being caught.