Remove the word "placeholder" from image checker
The Whitehall to Content Publisher migration code uses the presence of
the word "placeholder" to determine if we're using a default image.

Whitehall has recently been updated to remove "placeholder" as the
default for image alt text as this was an accessibility fail. The alt
text field should now be blank by default