Use GovukTest chrome options for Jasmine
GovukTest now provides a reusable configuration for headless chrome
which can support GOV.UK Docker running the application in containers as
a root user. Using these options provides an opt-in to this approach.

This also renames ChromeHeadlessJasmineConfigurer to
HeadlessChromeJasmineConfigurer so that headless chrome is used
consistently (as per GOV.UK test).